+++ src/helpers/copyToClipboard.js
@@ -1,0 +1,1 @@
+export const copyToClipboard = async copyText => navigator.clipboard.writeText(copyText);
