+++ src/helpers/regexForUrl.js
@@ -1,0 +1,2 @@
+export const regexForUrl =
+new RegExp(/^((http(s?)):\/\/)[-a-zA-Z0-9@:%._+~#=]{1,256}\.[a-zA-Z0-9()]{1,6}\b([-a-zA-Z0-9()@:%_+.~#?&//=]*)/);
